Attempt to use erlang enif_allc and friends in jq lib
There is a jv_alloc.c file that one should be able to change to change
the allocation and free function. Unfortunatly, it seems like the
jq code base is not consistent in how it uses the jv_alloc functions
and the stdlib function. I have found at least a couple of places
that uses strdup to allocate memory that is later freed with
jv_mem_free. This should be fixed upstream before we can attempt this
again.